As per code review: reorganized a bit + the NonNormalizedVersion now inherits from Version
